+++ .eslintrc.js
@@ -1,38 +1,39 @@
 module.exports = {
-    root: true,
-    "env": {
-        "node": true,
-        "commonjs": true,
-        "es6": true,
-        "jquery": false,
-        "jest": true,
-        "jasmine": true
-    },
-    "extends": "eslint:recommended",
-    "parserOptions": {
-        "sourceType": "module"
-    },
-    "rules": {
-        "indent": [
-            "warn",
-            "tab"
-        ],
-        "quotes": [
-            "warn",
-            "double"
-        ],
-        "semi": [
-            "error",
-            "always"
-        ],
-        "no-var": [
-            "error"
-        ],
-        "no-console": [
-            "error"
-        ],
-        "no-unused-vars": [
-            "warn"
-        ]
-    }
+	root: true,
+	"env": {
+		"node": true,
+		"commonjs": true,
+		"es6": true,
+		"jquery": false,
+		"jest": true,
+		"jasmine": true
+	},
+	"extends": "eslint:recommended",
+	"parserOptions": {
+		"sourceType": "module",
+		"ecmaVersion": "2018"
+	},
+	"rules": {
+		"indent": [
+			"warn",
+			"tab"
+		],
+		"quotes": [
+			"warn",
+			"double"
+		],
+		"semi": [
+			"error",
+			"always"
+		],
+		"no-var": [
+			"error"
+		],
+		"no-console": [
+			"error"
+		],
+		"no-unused-vars": [
+			"warn"
+		]
+	}
 };