+++ test/validator.spec.js
@@ -154,12 +154,13 @@ describe("Test add", () => {
 
 		check = v.compile(schema);
 
-		const context = {
+		const context = expect.objectContaining({
 			customs: expect.any(Object),
 			rules: expect.any(Array),
 			fn: expect.any(Array),
-			index: 2
-		};
+			index: 2,
+			async: false
+		});
 
 
 		expect(validFn).toHaveBeenCalledTimes(1);
@@ -445,7 +446,6 @@ describe("Test custom validation v1", () => {
 	});
 
 	it("should compile without error", () => {
-		const fn = jest.fn();
 		const check = v.compile({
 			num: {
 				type: "number",
@@ -453,7 +453,6 @@ describe("Test custom validation v1", () => {
 				max: 15,
 				integer: true,
 				custom(value){
-					fn(value);
 					if (value % 2 !== 0) return [{ type: "evenNumber", actual: value }];
 				}
 			}
@@ -471,7 +470,7 @@ describe("Test custom validation v1", () => {
 				max: 15,
 				integer: true,
 				custom(value){
-					fn(value);
+					fn(this, value);
 					if (value % 2 !== 0) return [{ type: "evenNumber", actual: value }];
 				}
 			}
@@ -479,7 +478,7 @@ describe("Test custom validation v1", () => {
 
 		const res = check({num: 12});
 		expect(res).toBe(true);
-		expect(fn).toBeCalledWith(12);
+		expect(fn).toBeCalledWith(v, 12);
 
 		expect(check({num: 8})[0].type).toEqual("numberMin");
 		expect(check({num: 18})[0].type).toEqual("numberMax");
@@ -496,7 +495,7 @@ describe("Test custom validation v1", () => {
 					fn(value);
 					if (value % 2 !== 0) return [{ type: "evenNumber", actual: value }];
 				}
-			}, 
+			},
 			b: {
 				type: "number",
 				custom(value){
@@ -533,7 +532,7 @@ describe("Test custom validation", () => {
 				max: 15,
 				integer: true,
 				custom(value, errors){
-					fn(value, errors);
+					fn(this ,value, errors);
 					if (value % 2 !== 0) errors.push({ type: "evenNumber", actual: value });
 					return value;
 				}
@@ -546,7 +545,7 @@ describe("Test custom validation", () => {
 	it("should work correctly with custom validator", () => {
 		const res = check({num: 12});
 		expect(res).toBe(true);
-		expect(fn).toBeCalledWith(12, []);
+		expect(fn).toBeCalledWith(v, 12, []);
 
 		expect(check({num: 8})[0].type).toEqual("numberMin");
 		expect(check({num: 18})[0].type).toEqual("numberMax");